Extract class for terminal command reader
